[wlan] We need modules to utilise the hack that sledges is proposing.
Most of the entry diffs are only to make features that used to be part
of the kernel remain part of the kernel, and not get inside an external
module. Being in a module caused unresolved symbols, probably the components
are not marked as dependencies inside the kernel anywhere.

The aim is to only have the wlan prima hardware module and everything else
can remain as it used to be - inside the kernel.

Change-Id: I244f7e9c64d1d1c1a404f10bbe6f7358edc1a2a1